+++ index.ts
@@ -1,6 +1,9 @@
-// FIZZ BUZZ
-import { FizzBuzz } from "./src/fizz-buzz/fizz-buzz";
+/* ------ FIZZ BUZZ -------*/
+// import { FizzBuzz } from "./src/fizz-buzz/fizz-buzz";
 
-for (let index = 1; index <= 100; index++) {
-    FizzBuzz.printFizzBuzzResult(index);
-}
+// for (let index = 1; index <= 100; index++) {
+//     FizzBuzz.printFizzBuzzResult(index);
+// }
+
+
+/* ------ TENNIS SCORE -------*/