+++ changelog.md
@@ -6,7 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 ## [Unreleased]
 
-- Tennis Score
+### Added
+
+- **Tennis Score**
+    - can Convert the number of exchanges won to the tennis score.
 
 ## [1.0.0] - 2024-06-27